# doc/OpenAPI草案.yaml
openapi: 3.0.3
info:
  title: 育英星宠 API
  version: 0.1.0
  description: >
    第一阶段 OpenAPI 草案。基于 React + Node.js + MySQL + WebSocket 架构，
    支持 Admin / Display 双端、固定分值积分规则、15 分钟 Display 解锁、即时领取兑换、
    AI 学情手动生成。
servers:
  - url: /api/v1
...
paths:
...
  /display/unlock:
    post:
      tags: [Display]
      summary: Display 进入教师操作模式
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DisplayUnlockRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DisplayUnlockResponse'
        '403':
          description: 无权解锁
  /display/unlock-status:
    get:
      tags: [Display]
      summary: 查询 Display 解锁状态
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: classId
          required: true
          schema:
            type: integer
            format: int64
        - in: query
          name: displayTerminalCode
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseResponse'
  /display/lock:
    post:
      tags: [Display]
      summary: Display 主动锁定
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseResponse'
  /display/entry-config:
    get:
      tags: [Display]
      summary: 读取进入页配置
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: classId
          required: false
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseResponse'
  /display/classes/{classId}/home:
    get:
      tags: [Display]
      summary: Display 班级主页聚合数据
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: classId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseResponse'
  /display/classes/{classId}/leaderboard:
    get:
      tags: [Display]
      summary: Display 排行榜数据
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: classId
          required: true
          schema:
            type: integer
            format: int64
        - in: query
          name: type
          required: true
          schema:
            type: string
            enum: [score, group, medal]
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseResponse'
  /display/classes/{classId}/reward-center:
    get:
      tags: [Display]
      summary: Display 兑换页数据
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: classId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseResponse'
...
  /score-rules:
    get:
      tags: [ScoreRules]
      summary: 积分规则列表
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: moduleType
          schema:
            type: string
            enum: [general, subject]
        - in: query
          name: subjectCode
          schema:
            type: string
        - in: query
          name: sceneCode
          schema:
            type: string
        - in: query
          name: displayEnabled
          schema:
            type: boolean
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseResponse'
    post:
      tags: [ScoreRules]
      summary: 新建积分规则
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScoreRuleUpsertRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseResponse'
  /score-rules/{id}:
    get:
      tags: [ScoreRules]
      summary: 积分规则详情
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseResponse'
    put:
      tags: [ScoreRules]
      summary: 更新积分规则
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScoreRuleUpsertRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseResponse'
  /score-records:
    get:
      tags: [ScoreRecords]
      summary: 积分记录查询
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: classId
          schema:
            type: integer
            format: int64
        - in: query
          name: studentId
          schema:
            type: integer
            format: int64
        - in: query
          name: subjectCode
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseResponse'
    post:
      tags: [ScoreRecords]
      summary: 单人加减分
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScoreRecordCreateRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseResponse'
  /score-records/batch:
    post:
      tags: [ScoreRecords]
      summary: 批量加减分
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScoreRecordBatchRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseResponse'
  /score-records/group:
    post:
      tags: [ScoreRecords]
      summary: 按组加减分
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScoreRecordGroupRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseResponse'
  /score-records/{id}/reverse:
    post:
      tags: [ScoreRecords]
      summary: 撤销或修正评价
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseResponse'
...
  /rewards:
    get:
      tags: [Rewards]
      summary: 奖励列表
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseResponse'
    post:
      tags: [Rewards]
      summary: 新建奖励
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RewardUpsertRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseResponse'
  /rewards/{id}:
    put:
      tags: [Rewards]
      summary: 更新奖励
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RewardUpsertRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseResponse'
  /reward-orders:
    get:
      tags: [Rewards]
      summary: 兑换记录查询
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: classId
          schema:
            type: integer
            format: int64
        - in: query
          name: studentId
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseResponse'
    post:
      tags: [Rewards]
      summary: 发起兑换
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RewardOrderCreateRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseResponse'
        '403':
          description: 任课教师无权兑换
        '409':
          description: 积分不足或库存不足
  /honors:
    get:
      tags: [Honors]
      summary: 荣誉列表
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseResponse'
  /honor-records:
    get:
      tags: [Honors]
      summary: 荣誉记录查询
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseResponse'
    post:
      tags: [Honors]
      summary: 发放荣誉
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HonorRecordCreateRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdPath:
      in: path
      name: id
      required: true
      schema:
        type: integer
        format: int64
  schemas:
    BaseResponse:
      type: object
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: ok
        data:
          nullable: true
...
    DisplayUnlockRequest:
      type: object
      required: [classId, displayTerminalCode]
      properties:
        classId:
          type: integer
          format: int64
        displayTerminalCode:
          type: string
    DisplayUnlockResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                unlockSessionId:
                  type: integer
                  format: int64
                expiredAt:
                  type: string
                  format: date-time
...
    ScoreRuleUpsertRequest:
      type: object
      required:
        [semesterId, moduleType, sceneCode, code, name, scoreType, scoreValue, sentiment]
      properties:
        semesterId:
          type: integer
          format: int64
        moduleType:
          type: string
          enum: [general, subject]
        subjectCode:
          type: string
        sceneCode:
          type: string
        code:
          type: string
        name:
          type: string
        scoreType:
          type: string
          enum: [add, deduct]
        scoreValue:
          type: integer
        dimension:
          type: string
        tag:
          type: string
        sentiment:
          type: string
          enum: [positive, negative]
        weight:
          type: number
          format: double
        aiSummaryText:
          type: string
        description:
          type: string
        isHighFrequency:
          type: boolean
        displayEnabled:
          type: boolean
        adminEnabled:
          type: boolean
    ScoreRecordCreateRequest:
      type: object
      required: [classId, studentId, ruleId, sourceTerminal]
      properties:
        classId:
          type: integer
          format: int64
        studentId:
          type: integer
          format: int64
        ruleId:
          type: integer
          format: int64
        remark:
          type: string
        sourceTerminal:
          type: string
          enum: [admin, display]
    ScoreRecordBatchRequest:
      type: object
      required: [classId, studentIds, ruleId, sourceTerminal]
      properties:
        classId:
          type: integer
          format: int64
        studentIds:
          type: array
          items:
            type: integer
            format: int64
        ruleId:
          type: integer
          format: int64
        remark:
          type: string
        sourceTerminal:
          type: string
          enum: [admin, display]
    ScoreRecordGroupRequest:
      type: object
      required: [classId, classGroupId, ruleId, sourceTerminal]
      properties:
        classId:
          type: integer
          format: int64
        classGroupId:
          type: integer
          format: int64
        ruleId:
          type: integer
          format: int64
        remark:
          type: string
        sourceTerminal:
          type: string
          enum: [admin, display]
...
    RewardUpsertRequest:
      type: object
      required: [code, name, scoreCost]
      properties:
        code:
          type: string
        name:
          type: string
        category:
          type: string
        imageUrl:
          type: string
        scoreCost:
          type: integer
        stockQty:
          type: integer
        isInfiniteStock:
          type: boolean
    RewardOrderCreateRequest:
      type: object
      required: [classId, studentId, rewardId, sourceTerminal]
      properties:
        classId:
          type: integer
          format: int64
        studentId:
          type: integer
          format: int64
        rewardId:
          type: integer
          format: int64
        sourceTerminal:
          type: string
          enum: [admin, display]
    HonorRecordCreateRequest:
      type: object
      required: [honorId, targetType, targetId, classId]
      properties:
        honorId:
          type: integer
          format: int64
        targetType:
          type: string
          enum: [student, class]
        targetId:
          type: integer
          format: int64
        classId:
          type: integer
          format: int64
        remark:
          type: string